ㅁfeat: Implement shopping cart functionality and UI enhancements

+++ backend/src/main/java/com/mateandgit/candestore/domain/product/dto/ProductResponse.java
@@ -2,11 +2,13 @@ package com.mateandgit.candestore.domain.product.dto;
 
 import com.mateandgit.candestore.domain.product.entity.Product;
 
+import java.math.BigDecimal;
+
 public record ProductResponse(
         Long id,
         String title,
         String description,
-        int price
+        BigDecimal price
 ) {
     public ProductResponse(Product product) {
         this(
